Narrow remaining CELT helper tables and math

+++ celt/dynalloc_test.go
@@ -152,7 +152,7 @@ func TestNoiseFloorComputation(t *testing.T) {
 				}
 				eMean := 0.0
 				if i < len(eMeans) {
-					eMean = eMeans[i]
+					eMean = float64(eMeans[i])
 				}
 
 				// libopus formula (float version):
@@ -256,7 +256,7 @@ func TestMaxDepthCalculation(t *testing.T) {
 				}
 				eMean := 0.0
 				if i < len(eMeans) {
-					eMean = eMeans[i]
+					eMean = float64(eMeans[i])
 				}
 				noiseFloor[i] = 0.0625*logNVal + 0.5 + float64(9-lsbDepth) - eMean + 0.0062*float64((i+5)*(i+5))
 			}
